+++ src/Traits/ColumnsValidationChecks.php
@@ -5,6 +5,7 @@ namespace Okipa\LaravelTable\Traits;
 use ErrorException;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Schema;
 use Okipa\LaravelTable\Column;
 
@@ -107,7 +108,7 @@ trait ColumnsValidationChecks
         $tableData = $this->tableData($column, $query);
         foreach ($searchedDatabaseColumns as $searchedDatabaseColumn) {
             if (! in_array($searchedDatabaseColumn, $tableData['columns'])) {
-                $dynamicMessagePart = ($tableAlias = array_get($tableData, 'alias'))
+                $dynamicMessagePart = ($tableAlias = Arr::get($tableData, 'alias'))
                     ? '« ' . $tableData['table'] . ' » (aliased as « ' . $tableAlias . ' ») table'
                     : '« ' . $tableData['table'] . ' » table';
                 $errorMessage = 'The table column with related « ' . $searchedDatabaseColumn . ' » database column is '